Separate broker time capabilities
Move Linux broker clock types into a dedicated time module and use a standalone time provider for lock tracing.


Copilot-Session: 239ec5f7-870a-4259-bcae-4ca85fb913a0
