+++ litebox_broker_platform_linux_userland/src/lib.rs
@@ -11,6 +11,8 @@
 
 mod socket;
 mod sync;
+mod time;
 
 pub use socket::LinuxSocketProvider;
-pub use sync::{LinuxSyncInstant, LinuxSyncPrimitivesProvider, LinuxSyncSystemTime};
+pub use sync::LinuxSyncPrimitivesProvider;
+pub use time::{LinuxInstant, LinuxSystemTime, LinuxTimeProvider};
