+++ litebox_broker_userland/src/linux.rs
@@ -14,7 +14,7 @@ use litebox_broker_core::socket::HOST_GATEWAY_IPV4_ADDRESS;
 use litebox_broker_core::{BrokerCore, BrokerCoreLimits, ObjectRights, PolicyEngine};
 use litebox_broker_platform_linux_userland::LinuxSocketProvider;
 #[cfg(feature = "lock_tracing")]
-use litebox_broker_platform_linux_userland::LinuxSyncPrimitivesProvider;
+use litebox_broker_platform_linux_userland::LinuxTimeProvider;
 use litebox_broker_protocol::message::{BrokerRequest, BrokerResponse};
 use litebox_broker_protocol::shared_buffer::SHARED_BUFFER_POOL_SIZE;
 use litebox_broker_transport::channel::HostReceive;
@@ -33,7 +33,7 @@ use super::{random::UserlandRandomProvider, stdio::UserlandStdioProvider};
 const PROXY_SHUTDOWN_TIMEOUT: Duration = Duration::from_secs(5);
 
 #[cfg(feature = "lock_tracing")]
-static LOCK_TRACING_PLATFORM: LinuxSyncPrimitivesProvider = LinuxSyncPrimitivesProvider;
+static LOCK_TRACING_PLATFORM: LinuxTimeProvider = LinuxTimeProvider;
 
 impl HostRequestSource for UnixControlRingHostRequestSource {
     fn recv_request(&mut self) -> IoResult<HostReceive<BrokerRequest>> {
